requirements.txt
pytest==8.0.0
allure-pytest==2.13.2
python-dotenv==1.1.1
requests==2.31.0
pydantic==1.10.14
typing-extensions==4.3.0
selene==2.0.0b17
selenium==4.16.0
webdriver-manager==3.8.6

jsonschema==4.22.0
